renamed CaseDetail to SubjectDetail; removed extraneous files
