+++ src/components/PropertySubsection/subjectDetailSubsection.js
